feat: implement UserAuth component for user authentication
